CLOCK has changed from uint32_t to uint64_t. Clkguard has been removed, which means no more large jumps backward in time to prevent CLOCK overflow, and all related complexity has been removed. Snapshot code has been updated to write and read the larger CLOCK values, which means all existing snapshots will no longer work.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@40432 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/types.h
@@ -46,7 +46,7 @@
 #  endif
 #endif
 
-typedef uint32_t CLOCK;
+typedef uint64_t CLOCK;
 
 /* Maximum value of a CLOCK.  */
 #undef CLOCK_MAX
